requirements.txt
asgiref==3.5.0
certifi==2022.6.15
chardet==5.1.0
charset-normalizer==2.1.0
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.3
#mod-wsgi==4.9.4
module-name==0.5.1
openpyxl==3.0.10
psycopg2==2.9.5
python-decouple==3.6
pytz==2021.3
requests==2.28.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.11
whitenoise==5.3.0
xmltodict==0.13.0
